fix(e2e): close deterministic world coverage gaps
