+++ .github/workflows/e2e-postgres.yml
@@ -57,6 +57,7 @@ jobs:
   e2e-worlds:
     name: e2e-worlds-postgres (${{ matrix.fixture_name }}, mock)
     runs-on: ubuntu-latest
+    timeout-minutes: 30
     needs: [changes, discover-fixtures]
     if: needs.changes.outputs.relevant == 'true'
     strategy:
