+++ .github/workflows/e2e-vercel.yml
@@ -60,6 +60,7 @@ jobs:
   e2e-worlds:
     name: e2e-worlds-vercel (${{ matrix.fixture_name }}, mock)
     runs-on: ubuntu-latest
+    timeout-minutes: 30
     needs: [changes, discover-fixtures]
     if: needs.changes.outputs.relevant == 'true'
     strategy:
